Experimental

+++ app/src/main/java/com/skepticalone/mecachecker/data/util/MoneyConverter.java
@@ -11,13 +11,13 @@ public final class MoneyConverter {
     private static final int MONEY_SCALE = 2;
 
     @TypeConverter
-    public static int moneyToCents(@NonNull BigDecimal money) {
-        return money.setScale(MONEY_SCALE, RoundingMode.HALF_UP).unscaledValue().intValue();
+    public static long moneyToCents(@NonNull BigDecimal money) {
+        return money.setScale(MONEY_SCALE, RoundingMode.HALF_UP).unscaledValue().longValue();
     }
 
     @TypeConverter
     @NonNull
-    public static BigDecimal centsToMoney(int cents) {
+    public static BigDecimal centsToMoney(long cents) {
         return BigDecimal.valueOf(cents, MONEY_SCALE);
     }
